+++ tools/perf.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 export NVC_LIBPATH=./lib
-./bin/nvc -a ../test/perf/$1.vhd && ./bin/nvc -e -V --native $1 && ./bin/nvc -r --stats $*
+./bin/nvc --native -a ../test/perf/$1.vhd -e -V $1 && ./bin/nvc -r --stats $*